# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Satpol PP",
        "description": "Dokumentasi ini berisi semua endpoint API yang tersedia untuk proyek Satpol PP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tugas": {
            "get": {
                "tags": [
                    "Tugas & Fungsi"
                ],
                "summary": "Menampilkan data Tugas & Fungsi",
                "description": "Mengambil data Tugas & Fungsi yang tersimpan (hanya ada satu).",
                "operationId": "getTugas",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tugas & Fungsi"
                ],
                "summary": "Membuat data Tugas & Fungsi baru",
                "description": "Hanya bisa dijalankan jika belum ada data sama sekali.",
                "operationId": "storeTugas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "konten"
                                ],
                                "properties": {
                                    "konten": {
                                        "type": "string",
                                        "example": "Ini adalah konten tugas dan fungsi."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    },
                    "409": {
                        "description": "Data sudah ada"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tugas/{id}": {
            "put": {
                "tags": [
                    "Tugas & Fungsi"
                ],
                "summary": "Memperbarui data Tugas & Fungsi",
                "description": "Memperbarui data Tugas & Fungsi yang sudah ada.",
                "operationId": "updateTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "konten"
                                ],
                                "properties": {
                                    "konten": {
                                        "type": "string",
                                        "example": "Ini adalah konten yang sudah diperbarui."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tugas & Fungsi"
                ],
                "summary": "Menghapus data Tugas & Fungsi",
                "operationId": "deleteTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}